requirements.txt
dash==2.18.2
plotly==5.24.1
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
